[frontend] refactor: выделение MushroomScreen (#6)
